Make spicedb a separate package
