+++ spicedb/__init__.py
@@ -1,5 +1,5 @@
 ################################################################################
-# spicedb.py
+# spicedb/__init__.py
 #
 # This set of routines handles the selection of SPICE kernels based on various
 # criteria related to body, instrument, time frame, etc. It also sorts selected
@@ -19,7 +19,7 @@ import textkernel
 
 import cspyce
 
-import sqlite_db as db
+import spicedb.sqlite_db as db
 
 # For testing and debugging
 DEBUG = False   # If true, no files are furnished.
